+++ article/article.go
@@ -38,50 +38,53 @@ func Parse(title, domain, article, references string) string {
 	formattedArticle := ""
 
 	for i, line := range lines {
-		isOnFirstOrLastLine := i == 0 || i == len(lines)-1
-		isOnReferences := line == "References"
-
-		if isOnFirstOrLastLine {
-			formattedArticle += normal + line + newLine
-
-			continue
-		}
-
-		if isOnReferences {
+		if line == "References" {
 			break
 		}
 
-		isOnQuote := isQuote(line)
-		isOnHeader := isHeader(lines, i, line)
-
-		if isOnQuote {
+		if isQuote(line) {
 			formattedArticle += normal + Faint(line).Italic().String() + newLine
+			formattedArticle = highlightReferencesInsideQuotes(formattedArticle)
 
 			continue
 		}
 
-		if isOnHeader {
+		if isHeader(lines, i, line) {
 			formattedArticle += normal + Bold(line).String() + newLine
 
 			continue
 		}
 
 		formattedArticle += normal + line + newLine
+		formattedArticle = highlightReferences(formattedArticle)
 	}
 
-	formattedArticle = highlightReferences(formattedArticle)
 	formattedReferences := formatReferences(references)
 
 	return wrappedTitle + formattedArticle + formattedReferences
 }
 
 func isHeader(lines []string, i int, line string) bool {
-	previousLine := lines[i-1]
+	var previousLine string
+	var nextLine string
+
+	isOnFirstLine := i == 0
+	isOnLastLine := len(lines) > i
+
+	if isOnFirstLine {
+		previousLine = ""
+	} else {
+		previousLine = lines[i-1]
+	}
+
+	if isOnLastLine {
+		nextLine = ""
+	} else {
+		nextLine = lines[i+1]
+	}
+
 	previousLineIsEmpty := len(previousLine) == 0
-
-	nextLine := lines[i+1]
 	nextLineLineIsEmpty := len(nextLine) == 0
-
 	currentLineIsNotIndented := !strings.HasPrefix(line, " ")
 
 	lineIsHeader := currentLineIsNotIndented && previousLineIsEmpty && nextLineLineIsEmpty
@@ -122,6 +125,29 @@ func highlightReferences(input string) string {
 	input = strings.ReplaceAll(input, "[14]", "["+White("14").String()+"]")
 	input = strings.ReplaceAll(input, "[15]", "["+Red("15").String()+"]")
 	input = strings.ReplaceAll(input, "[16]", "["+Yellow("16").String()+"]")
+
+	return input
+}
+
+func highlightReferencesInsideQuotes(input string) string {
+	faintItalic := "\033[2m" + "\033[3m"
+
+	input = strings.ReplaceAll(input, "[1]", "["+Red("1").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[2]", "["+Yellow("2").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[3]", "["+Green("3").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[4]", "["+Cyan("4").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[5]", "["+Blue("5").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[6]", "["+Magenta("6").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[7]", "["+White("7").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[8]", "["+BrightRed("8").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[9]", "["+BrightYellow("9").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[10]", "["+BrightGreen("10").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[11]", "["+BrightCyan("11").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[12]", "["+BrightBlue("12").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[13]", "["+BrightMagenta("13").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[14]", "["+White("14").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[15]", "["+Red("15").String()+faintItalic+"]")
+	input = strings.ReplaceAll(input, "[16]", "["+Yellow("16").String()+faintItalic+"]")
 
 	return input
 }